Conflicts:
	AgentensystemeAusarbeitung/log/log.log

+++ AgentensystemeAusarbeitung/src/data/MapAsArray.java
@@ -7,7 +7,7 @@ import org.apache.log4j.Logger;
 import org.apache.log4j.PropertyConfigurator;
 
 /**
- * Ich denke das wird die Bevorzugte variante die Karte dazustellen Unten Links
+ * Ich denke das wird die Bevorzugte variante die Karte darzustellen Unten Links
  * ist (0,0) Es wird immer von der Mitte aus begonnen die Karte zu befüllen Wenn
  * da nichts ist steht im Array Null Zu nutzen ist {@IMap} als Speicherort für
  * die Map, falls sich da was ändert. Dich sollte für die GUI nur getMap()
@@ -34,9 +34,9 @@ public class MapAsArray {
 	}
 
 	public Cord addNewField(Cell field, Cord cord) {
-		log.info("Add new field on: "+cord);
+		log.info("Add new field on: " + cord);
 		if (isInRange(cord)) {
-			resizeMap(10,cord);
+			resizeMap(10, cord);
 
 		}
 		if (map[cord.getX()][cord.getY()] == null) {
@@ -51,57 +51,57 @@ public class MapAsArray {
 	public Cord getCurrentLocation() {
 		return getMid();
 	}
-	
-	public Cell getCurrentField(Cord cord){
-		if(isInRange(cord)){
-			resizeMap(10,cord);
+
+	public Cell getCurrentField(Cord cord) {
+		if (isInRange(cord)) {
+			resizeMap(10, cord);
 		}
 		return map[cord.getX()][cord.getY()];
 	}
 
 	public List<Cord> getNeighbours(Cord cord) {
-		log.info("GetNeighbours on: "+cord);
+		log.info("GetNeighbours on: " + cord);
 		List<Cord> list = new ArrayList<>();
-		Cord test = new Cord(cord.getX()-1, cord.getY());
-		if(isInRange(test)){
+		Cord test = new Cord(cord.getX() - 1, cord.getY());
+		if (isInRange(test)) {
 			resizeMap(10, cord);
 			return getNeighbours(cord);
 		}
 		list.add((map[test.getX()][test.getY()] == null) ? test : null);
-		test = new Cord(cord.getX(), cord.getY()-1);
-		if(isInRange(test)){
+		test = new Cord(cord.getX(), cord.getY() - 1);
+		if (isInRange(test)) {
 			resizeMap(10, cord);
 			return getNeighbours(cord);
 		}
 		list.add((map[test.getX()][test.getY()] == null) ? test : null);
-		test = new Cord(cord.getX(), cord.getY()+1);
-		if(isInRange(test)){
+		test = new Cord(cord.getX(), cord.getY() + 1);
+		if (isInRange(test)) {
 			resizeMap(10, cord);
 			return getNeighbours(cord);
 		}
 		list.add((map[test.getX()][test.getY()] == null) ? test : null);
-		test = new Cord(cord.getX()+1, cord.getY());
-		if(isInRange(test)){
+		test = new Cord(cord.getX() + 1, cord.getY());
+		if (isInRange(test)) {
 			resizeMap(10, cord);
 			return getNeighbours(cord);
 		}
 		list.add((map[test.getX()][test.getY()] == null) ? test : null);
-		log.info("Current Location after analysis of neighbours: "+cord);
+		log.info("Current Location after analysis of neighbours: " + cord);
 		return list;
 	}
 
 	public int getFieldIndex(Cord cord) {
-		log.info("analsysis the fieldindex: "+cord);
+		log.info("analsysis the fieldindex: " + cord);
 		int index = 0;
 		for (int i = -1; i <= 1; ++i) {
 			for (int j = -1; j < 1; ++j) {
-				Cord cordT = new Cord(cord.getX()+i, cord.getY()+j);
-				if(!isInRange(cordT)){
+				Cord cordT = new Cord(cord.getX() + i, cord.getY() + j);
+				if (!isInRange(cordT)) {
 					index += (map[cordT.getX()][cordT.getY()] != null) ? 1 : 0;
 				}
 			}
 		}
-		log.info("currentLocation after analysis of fieldindex: "+cord);
+		log.info("currentLocation after analysis of fieldindex: " + cord);
 		return index;
 	}
 
@@ -110,7 +110,7 @@ public class MapAsArray {
 	}
 
 	private void resizeMap(int resize, Cord cord) {
-		log.info("resize Map with cord: "+cord);
+		log.info("resize Map with cord: " + cord);
 		Cell[][] newMap = new Cell[map.length + resize][map.length + resize];
 		for (int i = 0; i < map.length; i++) {
 			for (int j = 0; j < map.length; j++) {
@@ -120,36 +120,37 @@ public class MapAsArray {
 		map = newMap;
 		cord.setX(cord.getX() + resize / 2);
 		cord.setY(cord.getY() + resize / 2);
-		log.info("resized cord: "+cord);
-		
+		log.info("resized cord: " + cord);
+
 	}
 
 	public Cord getMid() {
-		return new Cord((int) (map.length / 2), (int) (map[0].length / 2));
+		return new Cord(map.length / 2, map[0].length / 2);
 	}
 
 	private boolean isInRange(Cord c) {
-		return !(c.getX() < map.length-1 && c.getY() < map[0].length-1 && c.getX() >= 0 && c.getY() >= 0);
+		return !(c.getX() < map.length - 1 && c.getY() < map[0].length - 1 && c.getX() >= 0 && c.getY() >= 0);
 	}
 
 	public void print(Cell[][] map) {
 		StringBuilder b = new StringBuilder();
-		for (int i = 0 ; i < map.length; i++) {
-			for (int j = 0 ; j < map[i].length ; j++) {
+		for (int i = 0; i < map.length; i++) {
+			for (int j = 0; j < map[i].length; j++) {
 				Cell cell = map[j][i];
 				if (cell == null) {
 					b.append("-");
-				}else if(cell.getStench()>0){
+				} else if (cell.getStench() > 0) {
 					b.append("S");
-				}else if(cell.isRock()){
+				} else if (cell.isRock()) {
 					b.append("R");
-				}else if(i==getMid().getX() && j == getMid().getY() || j==getMid().getX() && i == getMid().getY()){
+				} else if (i == getMid().getX() && j == getMid().getY()
+						|| j == getMid().getX() && i == getMid().getY()) {
 					b.append("H");
-				}else if(cell.getFood()>0){
+				} else if (cell.getFood() > 0) {
 					b.append("F");
-				}else if(cell.isTrap()){
+				} else if (cell.isTrap()) {
 					b.append("T");
-				}else{
+				} else {
 					b.append("O");
 				}
 			}
